==> 0_Estrazione_Dati_Tesi_Da_File_JSON_MongoDB.ipynb <==
Progetto: **Classificazione di testi descrittivi per destinazione d'uso, inerenti ai beni sottoposti ad aste giudiziarie italiane.**

Studente: **Alessandro Monolo** | *10439147*

Relatore: Marco Brambilla

Referente aziendale: Simone Redaelli

Master: Data Science & Artificial Intelligence

Università: Politecnico di Milano

<hr style="border:1px solid black">

# Filtro file JSON estratto da MongoDB con le colonne inerenti al progetto

- **1.** **Filtro** il data frame estratto da MongoDB con tutte le colonne che mi servono;


- **2.** **Analizzo** il data frame;


- **3.** **Salvo** il data frame filtrato e corretto in un fomato **CSV**.

<hr style="border:1px solid black">

**Importo le librerie che mi servono:**

import pandas as pd

#### Set pandas options:

pd.set_option('display.max_colwidth', None)
pd.options.display.max_rows = 5000
pd.options.display.max_columns = 1000
pd.options.mode.chained_assignment = None 

**Importo file JSON**

df_json = pd.read_json('D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\0_Dataset_MongoDB\\10.08.2023\\Descrizioni beni DD2.json')

### 1 - Seleziono le colonne con le quali mi serve lavorare

column_list = ["DestinazioneUso",
               "CategoriaCatastale",
               "Provincia",
               "Comune",
               "Tribunale",
               "NumLotto",
               "Superficie",
               "NumeroVani",
               "TipoCatasto",
               "Descrizione"]

**Filtro il Data Frame Pandas**

df = df_json.loc[:, column_list]

### 2 - Controllo il df estratto appena filtrato

df.shape

- Il Data Frame è composto da **10** colonne e **251.510** righe;

df.info()

- Attualmente **nessuna** colonna è espressa in valori **numerici**;
- Tutte le colonne sono quindi **Object** Format.

print("Somma delle Null Values per singola colonna del data frame:")
print(df.isnull().sum())
print("\n")
print("Null Values espresse in termini percentuali per singola colonna del data frame:")
print(df.isnull().sum()*100 / len(df))

- La maggior parte delle colonne non presenta Null Values al loro interno;
- Le uniche sono le colonne "**Superficie**" e "**NumeroVani**".
    - Questa presenza massiccia di Null Values può essere inerente al fatto che all'interno del data frame i suddetti campi vengono/possono essere mappati solo nel caso in cui non si tratti di terreni (ovvero beni con destinazione d'uso equivalente a LAND);
- Il campo "**Descrizione**" è vuoto in 38 casi;
    - Devo quindi rimuovere tutte quelle righe dove in questo campo non esistono stringhe descrittive del bene, prima di esportare il data frame appena filtrato.

df.describe()

- Ci sono **8** classi nel campo **destinazione d'uso**;
    - La più frequente è la classe "**RESIDENTIAL**", presente **175.185** volte, con categoria catastale **A2**;
- Il comune più citato è il comune di ROMA;
- Mentre il tribunale di assegnazione più frequente nel data frame è quello di **Bergamo**;
- Da notare inoltre che la Mq minima è pari a zero, quindi un ipotetico errore da indagare successivamente.

**Droppo tutte le righe dove in entrambi i campi Descrizione e DescrizioneLotto i valori sono assenti**

df_filtrato = df.dropna(subset=['Descrizione'], how='all')

**Stampo ora il totale delle righe droppate**

print("Numero righe rimosse: ", df.shape[0] - df_filtrato.shape[0])

- Sono stati eliminati **38 beni** dove nel campo Descrizione era effettivamente presente una descrizione;

### Elimino tutti i beni che non ha una Destinazione D'Uso, ovvero tutti quei beni che sono identificati come "OTHER" nella Target Variable.

df_filtrato = df_filtrato[df_filtrato["DestinazioneUso"] != "OTHER"]

#### Prima di salvare il nuovo data frame filtrato, sostituisco la stringa "NA" con "NAP" per la provincia di Napoli, al fine di evitare che durante l'import nei prossimi step possa essere letta come una Null Values

# Sostituisco la stringa NA con NAP per evitare null values:
df_filtrato.loc[df_filtrato["Provincia"] == "NA", "Provincia"] = "NAP"

### 3 - Procedo a salvare il data frame filtrato in un file formato CSV: 

df_filtrato_csv = df_filtrato.to_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\1_Dataset_Filtrato\\Descrizioni_Filtrate_10.08.2023.csv",
                                     index=False)
