==> 4_Find_&_Drop_Bad_Data_Dal_Data_Frame.ipynb <==
Progetto: **Classificazione di testi descrittivi per destinazione d'uso, inerenti ai beni sottoposti ad aste giudiziarie italiane.**

Studente: **Alessandro Monolo** | *10439147*

Relatore: Marco Brambilla

Referente aziendale: Simone Redaelli

Master: Data Science & Artificial Intelligence

Università: Politecnico di Milano

<hr style="border:1px solid black">

## Find & Drop Bad Data per ogni colonna del data frame e dei duplicated values rimasti.

- **1.** Analizzo i valori presenti in ogni campo del data frame;


- **2.** Filtro il data frame da tutti i valori che rappresentano valori eccezionali al suo interno;


- **3.** Trovo ulteriori valori duplicati e li elimino dal mio data frame;


- **4.** Conclusioni;


- **5.** Export del data frame in file CSV.

<hr style="border:1px solid black">

**Importo le librerie che mi servono:**

import pandas as pd
import numpy as np
import re

#### Set pandas options:

pd.set_option('display.max_colwidth', None)
pd.options.display.max_rows = 5000
pd.options.display.max_columns = 1000
pd.options.display.float_format = '{:.2f}'.format

**Importo file CSV**

df = pd.read_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\4_Dataset_Bad_Data\\Dataset_Bad_Data.csv",
                 dtype={'Numero_Lotto' : 'Int64'})

### 1 - Analizzo i campi del data frame e quantifico Bad Data o valori eccezionali, in ognuno di essi

- Dopo aver esaminato le unique values presenti nei campi del data frame, conteggio i valori che presentano delle eccezioni o degli errori all'interno dei vari campi; qui di seguito l'esamina emersa.

print("1 - Numero di beni che presentano un numero lotto superiore a 1000:", '\033[1m','\033[15;31;43m',
      (df['Numero_Lotto'] > 1000).sum(), '\033[0m')

print("2 - Numero di beni con categoria RESIDENTIAL che contano più di 100 vani:", '\033[1m','\033[15;31;43m',
      ((df['Destinazione_Uso'] == "RESIDENTIAL") & (df['Numero_Vani'] > 100)).sum(), '\033[0m')

print("3 - Numero di beni con categoria diversa da LAND con superficie maggiore di 10.000 Mq:", '\033[1m','\033[15;31;43m',
      ((df['Destinazione_Uso'] != "LAND") & (df['Superficie'] > 10000)).sum(), '\033[0m')

print("4 - Numero di beni con categoria diversa da STORAGE con superficie inferiore ad 1 Mq:", '\033[1m','\033[15;31;43m',
      ((df['Destinazione_Uso'] != "STORAGE") & (df['Superficie'] < 1)).sum(), '\033[0m')

print("5 - Numero di beni che presentano una descrizione con più di 3000 caratteri:", '\033[1m','\033[15;31;43m',
      (df['N_Caratteri'] > 3000).sum(), '\033[0m')

### 2 - Filtro il Data Frame dai valori emersi precedentemente

- Per ogni campo del data frame dove sono emersi Bad data, filtro il data frame finale

# Conto i beni rimasti prima del drop dei Bad data:
print("Numero beni prima del drop dei Bad Data:", '\033[1m','\033[15;31;43m', df.shape[0], '\033[0m')

# 2.1 - Elimino i beni con Numero lotto superiore a 1000:
df = df[~((df['Numero_Lotto'] > 1000) & (~df['Numero_Lotto'].isnull()))]
# 2.2 - Elimino i beni con numero vani maggiori di 100 con destunazione Residential:
df = df[~(((df['Destinazione_Uso'] == "RESIDENTIAL") & (df['Numero_Vani'] > 100)) & (~df['Numero_Vani'].isnull()))]
# 2.3 - Elimino i beni con superficie maggiore di 10.000 Mq per i beni con destinazione diversa da Land:
df = df[~(((df['Destinazione_Uso'] != "LAND") & (df['Superficie'] > 10000)) & (~df['Superficie'].isnull()))]
# 2.4 - Elimino i beni con superficie minore di 1 Mq con destinazione diversa da Storage:
df = df[~(((df['Destinazione_Uso'] != "STORAGE") & (df['Superficie'] < 1)) & (~df['Superficie'].isnull()))]
# 2.5 - Elimino i beni con una lunghezza della descrizione del bene maggiore di 3000 caratteri:
df = df[~((df['N_Caratteri'] > 3000) & (~df['N_Caratteri'].isnull()))]

# Conto i beni rimasti dopo il drop dei Bad data:
print("Numero beni dopo il drop dei Bad Data:", '\033[1m','\033[15;31;43m', df.shape[0], '\033[0m')
# Conteggio dei beni droppati:
print("Numero dei beni eliminati", '\033[1m','\033[15;31;43m', 194553 - 193245 , '\033[0m')

### 3 - Trovo ed elimino ulteriori beni duplicati dal data frame:

# Quantifico i beni duplicati ancora presenti nel mio data frame:
duplicated_values = len(df[df.duplicated() ])
# Mostro il valore ottenuto:
print("Attualmente ci sono ancora", '\033[1m', '\033[15;31;43m', duplicated_values, '\033[0m', "beni duplicati nel data frame.")

# Mostro il numero di beni presenti in questo momento nel data frame:
print("Numero beni prima del drop dei beni duplicati rimasti:", '\033[1m','\033[15;31;43m', df.shape[0], '\033[0m')
# Elimino i beni duplicati dal data frame:
df = df.drop_duplicates(keep='first')

# Mostro quanti beni rimasti sono nel data frame senza i beni duplicati:
print("Numero beni dopo il drop dei beni duplicati rimasti:", '\033[1m','\033[15;31;43m', df.shape[0], '\033[0m')
# Mostro la differenza ottenuta da prima a dopo:
print("Numero dei beni eliminati", '\033[1m','\033[15;31;43m', 193245-193181 , '\033[0m')

### 4 - Conclusioni:

- Sono stati individuati ed eliminati alcuni **Bad Data** presenti nelle colonne numeriche el data frame.
    - In particolare sono stati eliminati **1.308 beni** che rappresentavano per i relativi campi, valori eccezionali o probabilmente sbagliati.Si tratta in questo caso del **0.7 % del totale dei beni iniziali**.
- Successivamente sono stati individuati ulteriori **beni duplicati** all'interno del data frame.
    - In particolare sono stati eliminati **64** beni duplicati dal data frame. Trattandosi quindi del **0.03% del totale dei beni rimasti**.

### 5 - Export data frame file CSV:

df_csv = df.to_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\5_Dataset_FillNullValues_&_DropNa\\Dataset_FillNullValues_&_DropNa.csv",
                   index=False)
