==> 7.2_Pre_Processing_Per_Random_Forest.ipynb <==
Progetto: **Classificazione di testi descrittivi per destinazione d'uso, inerenti ai beni sottoposti ad aste giudiziarie italiane.**

Studente: **Alessandro Monolo** | *10439147*

Relatore: Marco Brambilla

Referente aziendale: Simone Redaelli

Master: Data Science & Artificial Intelligence

Università: Politecnico di Milano

<hr style="border:1px solid black">

## Pre-Processing per data frame da usare nel modello predittivo "Random Forest"

- **1.** Creo una **nuova colonna** denominata **"Regione"** basata sul campo "Provincia";


- **2.** **Creo** una **nuova colonna**, basata su quella **booleana** **"Catasto_Fabbricati"**, ora in formato **Integer**;


- **3.** Fase di **Pre-Processing** alla **target variable "Destinazione d'uso"**, utilizzando la funzione **"LabelEncoder"**;


- **4.** Fase di **Pre-Processign** per le **variabili categoriche** **"Tribunale", "Provincia" & "Regione"**, applicando **LabelEncoder**;


- **5.** **Droppo** le colonne **non utili al modello**;


- **6.** **Conclusioni**;


- **7.** **Export** data frame in formato **csv**.


<hr style="border:1px solid black">

**Importo le librerie che mi servono:**

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import math
import warnings
from matplotlib import cm
import string
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.preprocessing import OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import LabelEncoder

#### Set pandas options:

pd.set_option('display.max_colwidth', None)
pd.options.display.max_rows = 5000
pd.options.display.max_columns = 1000
pd.options.display.float_format = '{:.6f}'.format
pd.options.mode.chained_assignment = None

**Importo file CSV**

df = pd.read_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\8_Dataset_Benchmark\\Dataset_Benchmark.csv",
                 dtype={'Numero_Lotto' : 'Int64'})

### 1 - Creo una nuova colonna "Regione", mappandola in base alle provincie di riferimento

# Creo un dizionario con tutte le provincie d'Italia presenti nel campo "Provincia" del mio data frame:
province_to_region = {
    'TO': 'Piemonte',
    'TE': 'Abruzzo',
    'MT': 'Basilicata',
    'AT': 'Piemonte',
    'RC': 'Calabria',
    'FI': 'Toscana',
    'VT': 'Lazio',
    'RM': 'Lazio',
    'AL': 'Piemonte',
    'BI': 'Piemonte',
    'ME': 'Sicilia',
    'IM': 'Liguria',
    'CS': 'Calabria',
    'VC': 'Piemonte',
    'SV': 'Liguria',
    'CN': 'Piemonte',
    'FG': 'Puglia',
    'SP': 'Liguria',
    'NO': 'Piemonte',
    'LI': 'Toscana',
    'MB': 'Lombardia',
    'BA': 'Puglia',
    'MI': 'Lombardia',
    'VA': 'Lombardia',
    'AO': 'Valle d\'Aosta',
    'OR': 'Sardegna',
    'VB': 'Piemonte',
    'PV': 'Lombardia',
    'CZ': 'Calabria',
    'SU': 'Sardegna',
    'GE': 'Liguria',
    'PC': 'Emilia-Romagna',
    'SA': 'Campania',
    'CO': 'Lombardia',
    'SR': 'Sicilia',
    'OT': 'Sardegna',
    'SS': 'Sardegna',
    'GR': 'Toscana',
    'SI': 'Toscana',
    'PR': 'Emilia-Romagna',
    'CL': 'Sicilia',
    'PI': 'Toscana',
    'LT': 'Lazio',
    'RG': 'Sicilia',
    'MS': 'Toscana',
    'SO': 'Lombardia',
    'AV': 'Campania',
    'AR': 'Toscana',
    'BG': 'Lombardia',
    'TV': 'Veneto',
    'LO': 'Lombardia',
    'BL': 'Veneto',
    'LC': 'Lombardia',
    'CR': 'Lombardia',
    'PN': 'Friuli-Venezia Giulia',
    'CT': 'Sicilia',
    'UD': 'Friuli-Venezia Giulia',
    'EN': 'Sicilia',
    'VR': 'Veneto',
    'PA': 'Sicilia',
    'VE': 'Veneto',
    'BS': 'Lombardia',
    'RE': 'Emilia-Romagna',
    'RN': 'Emilia-Romagna',
    'FE': 'Emilia-Romagna',
    'RA': 'Emilia-Romagna',
    'PD': 'Veneto',
    'NU': 'Sardegna',
    'MN': 'Lombardia',
    'VI': 'Veneto',
    'RO': 'Veneto',
    'BN': 'Campania',
    'VV': 'Calabria',
    'MO': 'Emilia-Romagna',
    'BZ': 'Trentino-Alto Adige',
    'KR': 'Calabria',
    'TR': 'Umbria',
    'BO': 'Emilia-Romagna',
    'GO': 'Friuli-Venezia Giulia',
    'TS': 'Friuli-Venezia Giulia',
    'LE': 'Puglia',
    'OG': 'Sardegna',
    'FC': 'Emilia-Romagna',
    'IS': 'Molise',
    'PU': 'Marche',
    'AN': 'Marche',
    'FM': 'Marche',
    'MC': 'Marche',
    'AP': 'Marche',
    'PE': 'Abruzzo',
    'LU': 'Toscana',
    'AQ': 'Abruzzo',
    'CE': 'Campania',
    'PT': 'Toscana',
    'PO': 'Toscana',
    'TP': 'Sicilia',
    'NAP': 'Campania',
    'PG': 'Umbria',
    'CH': 'Abruzzo',
    'RI': 'Lazio',
    'FR': 'Lazio',
    'TA': 'Puglia',
    'CB': 'Molise',
    'BR': 'Puglia',
    'BT': 'Puglia',
    'PZ': 'Basilicata',
    'AG': 'Sicilia',
    'CA': 'Sardegna',
    'VS': 'Sardegna'
}

# Mappo una nuova colonna "Regione", con il dizionario creato sopra:
df['Regione'] = df['Provincia'].map(province_to_region)

### 2 - Trasformo la colonna Booleana "Fabbricato_Catasto", in una colonna Integer

# Creo una nuova colonna basata sulla colonna Catasto_Fabbricati, assegnando per ogni True = 1 e per ogni False = 0:
df['Catasto_Fabbricati_Int'] = df['Catasto_Fabbricati'].astype(int)

### 3 - Fase di Pre-Processign per la Target Variable "Destinazione d'uso", applicando LabelEncoder

- Per la fase di **Pre-Processing** inerente alla target variable, utilizzo la libreria **sci-kit learn** e uso la funzione **"LabelEncoder"**, utile a trasformare ogni categoria presente nella target variable in un **integer progressivo**, da 0 a N.

# Instanzio in una nuova variabile il modello per l'encoding della target variable:
label_encoder = LabelEncoder()

# Applico il modello di encoding alla mia target variable e salvo il risultato in una nuova colonna:
df['Destinazione_Uso_Encoded'] = label_encoder.fit_transform(df['Destinazione_Uso'])

### 4 - Fase di Pre-Processign per le variabili categoriche "Tribunale", "Provincia" & "Regione", applicando LabelEncoder

# Applico il modello di encoding alla mia target variable e salvo il risultato in una nuova colonna:
df['Tribunale_Encoded'] = label_encoder.fit_transform(df['Tribunale'])

# Applico il modello di encoding alla mia target variable e salvo il risultato in una nuova colonna:
df['Provincia_Encoded'] = label_encoder.fit_transform(df['Provincia'])

# Applico il modello di encoding alla mia target variable e salvo il risultato in una nuova colonna:
df['Regione_Encoded'] = label_encoder.fit_transform(df['Regione'])

### 5 - Elimino le colonne che non mi servono per il modello predittivo

# Listo le colonne che voglio droppare dal data frame:
columns_to_drop = ["Destinazione_Uso",
                   "Comune",
                   "Categoria_Catastale",
                   "Descrizione_Bene",
                   "Clean_Descrizione_Bene",
                   "Words",
                   "Catasto_Fabbricati",
                   "Regione",
                   "Provincia",
                   "Tribunale"]

# Droppo le colonne listate:
df = df.drop(columns=columns_to_drop)

### 6 - Conclusioni

for index, (col, dtype) in enumerate(df.dtypes.items(), 1):
    print(f'N° Colonna: {index} - Nome Colonna: {col} - Tipo di dato: {dtype}\n________________________________________________________________________________\n')

# Numero di null values per colonna del data frame:
df.isnull().sum()

# Composizione del data frame ottenuto:
df.shape

# Composizione delle prime 10 righe di ogni colonna nel data frame:
df.head(10)

# Conto il numero di integers e float nel data frame:
integer_columns = df.select_dtypes(include=['int']).columns
num_integer_columns = len(integer_columns)
print("Numero totale di colonne in formato Integer:", num_integer_columns)

float_columns = df.select_dtypes(include=['float64']).columns
num_float_columns = len(float_columns)
print("Numero totale di colonne in formato float:", num_float_columns)

- Ogni variabile categoria presente nel data frame "Regione", "Provincia" & "Tribunale", è stata trasfornata attraverso la funzione di pre-processing "LabelEncoder", ottenendo un numero progressivo da 0 a N per ogni colonna:
    - In particolare nella colonna "Regione", sono presenti integers progressivi da 0 a 19;
    - Nella colonna **"Tribunale"**, sono presenti **integers** progressivi da **0 a 186**;
    - Nella colonna **"Provincia"**, sono presenti **integers** progressivi da **0 a 108**.
    

- Non ci sono **Null Values** all'interno delle colonne del data frame;


- Non ci sono **Object columns** nela data frame dopo l'aplicazione di "One-Hot Encoder";


- Il data frame è composto quindi da **166.040 righe** e **10 colonne**;


- In totale ci sono **7 Integer columns** e **3 Float columns** nel data frame;


- Dopo aver creato le derivanti dummies variable, le **Object columns sono state eliminate** dal data frame in quanto il modello **Random Forest** necessità solo di variabili numeriche;


- Anche l'unica variabile inizialmente **Booleana** è stata eliminata dopo essere stata trasformata in una variabile 0 / 1.

### 8 - Export data frame in formato CSV:

df_csv = df.to_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\9_Dataset_Random_Forest\\Dataset_Random_Forest.csv",
                   index=False)
