==> 9.5_FastText_&_Word2Vec_Tokenizer_Text_PreProcess.ipynb <==
Progetto: **Classificazione di testi descrittivi per destinazione d'uso, inerenti ai beni sottoposti ad aste giudiziarie italiane.**

Studente: **Alessandro Monolo** | *10439147*

Relatore: Marco Brambilla

Referente aziendale: Simone Redaelli

Master: Data Science & Artificial Intelligence

Università: Politecnico di Milano

<hr style="border:1px solid black">

## Fase di Preprocessing per il modello FastText e per i benchmark eseguiti con Word2Vec Tokenizer


- **1.** **<u>Pre-Processing sulla variabile testuale:</u>**
    - Per la fase di pre-processing sulla variabile testuale invece, ho eseguito i seguenti passaggi:
        - **1.1** **Sostituisco tutti caratteri che non sono caratteri di parole, spazi, o singoli apici con spazi**;
        - **1.2** **Elimino tutti i numeri presenti nella colonna testuale per FastText**;
        - **1.3** **Elimino dopo aver controllato tutte le parole che sono sotto i due caratteri**;
        - **1.4** **Elimino tutti i doppi spazi**;
        

- **2.** **<u>Filtro il data frame</u>**;


- **3.** **<u>Conclusioni</u>**


- **4.** **<u>Export final CSV</u>**


<hr style="border:1px solid black">

**Importo le librerie che mi servono:**

import pandas as pd
import numpy as np
import re
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
from matplotlib import cm
from PIL import Image
import string
import unicodedata
from sklearn.model_selection import train_test_split
import nltk
from nltk.stem import WordNetLemmatizer
from nltk.corpus import stopwords
import fasttext

#### Set pandas options:

pd.set_option('display.max_colwidth', None)
pd.options.display.max_rows = 5000
pd.options.display.max_columns = 1000
pd.options.display.float_format = '{:.4f}'.format
pd.options.mode.chained_assignment = None

**Importo file CSV**

df = pd.read_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\11_Dataset_TFIDF_LR_&_SVM\\Dataset_TFIDF_LR_&_SVM.csv")

### 1 - Concludo la fase di preprocessing per il modello FastText:

- **1.1 Sostituisco tutti caratteri che non sono caratteri di parole, spazi, o singoli apici con spazi:**

def remove_words_with_cypher(text):
    return ' '.join(word for word in text.split() if not re.search(r'\w*\d\w*', word))

# Apply the function to the column
df["Descrizione_Bene_FastText"] = df["Descrizione_Bene_Lemm_Spacy"].apply(remove_words_with_cypher)

- **1.2 Elimino tutti i numeri presenti nella colonna testuale per FastText:**

def preprocessor(text):
    text = re.sub(r'[^\w\s\']',' ', text)
    return text

df["Descrizione_Bene_FastText"] = df["Descrizione_Bene_FastText"].apply(preprocessor)

- **1.3 Elimino dopo aver controllato tutte le parole che sono sotto i due caratteri:**

# Definisco una user defined function che sia in grado di eliminare tutte le parole composte da un singolo carattere:
def filter_single_char_words(text):
    words = text.split()
    filtered_words = [word for word in words if len(word) > 1]
    return ' '.join(filtered_words)

# Applico la user defined function alla colonna testuale pre-processata:
df['Descrizione_Bene_FastText'] = df['Descrizione_Bene_FastText'].apply(filter_single_char_words)

- **1.4 Elimino tutti i doppi spazi:**

df["Descrizione_Bene_FastText"] = df["Descrizione_Bene_FastText"].apply(lambda x: ' '.join(x.split()))

### 2 - Filtro il data frame con solo le colonne che mi servono per la text classification

# Seleziono le colonne da eliminare
columns_to_drop = ["Descrizione_Bene_Pre_Process",
                   "Descrizione_Bene_Lemm_Spacy"]

# Mantengo solo le colonne che mi servono per la text classification:
df = df.drop(columns=columns_to_drop)

### 3 - Conclusioni

- In questa **ultima fase di preprocessing** ho eliminato, partendo dalla variabile testuale pulita, preprocessata e lemmatizzata:
    - Tutti i numeri;
    - I caratteri ambigui;
    - Doppi spazi;
    - Stringhe con meno di due caratteri;
- Ho quindi ridotto la variabile testuale a circa **39K** stringhe univoche contate.
- Di conseguenza, la variabile ora pronta per il modello supervisionato **FastText**, risulta essere più scarna e pulita rispetto a quelle utlizzate con TFIDF e BoW.

### 4 - Export CSV Finale

df_csv = df.to_csv("D:\\Master_Cefriel_DS_AI_Monolo\\0_Project_Work\\Dataset\\13_Dataset_FastText\\Dataset_FastText.csv", 
                   index=False)
